+++ app/models/competition.rb
@@ -17,7 +17,9 @@ class Competition < ApplicationRecord
   ## updater
   def normalize
     year = self.start_date.year
-    country_city = country || city.to_s.upcase.gsub(/\s+/, '_')        
+    country_city = country || city.to_s.upcase.gsub(/\s+/, '_')
+
+=begin
     ary = case name
           when /Grand Prix .*Final/, /^ISU GP.*Final/
             [:isu, :gp, "GPF#{year}"]
@@ -53,10 +55,26 @@ class Competition < ApplicationRecord
           else
             [:unknown, :unknown, name.to_s.gsub(/\s+/, '_')]
           end
-    self.competition_class ||= ary[0]
-    self.competition_type ||= ary[1]
-    self.short_name ||= ary[2]
-    self.name = ary[3] if ary[3]
+=end
+
+    ary = nil
+    data = YAML::load_file(File.join(Rails.root.join('config'), 'competition_normalize.yml'))
+    data.each do |key, value|
+      if name =~ Regexp.new(key)
+        ary = value
+        break
+      end
+    end
+
+    if ary.nil?
+      ary = [:unknow, :unknow, name.to_s.gsub(/\s+/, '_')]
+    end
+    hash = { year: year, city: city, country_city: country_city }
+    
+    self.competition_class ||= ary[0].to_sym
+    self.competition_type ||= ary[1].to_sym
+    self.short_name ||= ary[2] % hash
+    self.name = ary[3] % hash if ary[3]
     
     self
   end